64 bytes STR _PARAM
+ rewritten network connection params to 64 bytes as command(qint32 com, STR_PARAM p1, STR_PARAM p2, qint32 time)
+ expanded the command table in database for types and comments of params
***** need to struct data in database signals and slots next
***** think about an object manager for models and especialy for formats in the own window

+++ STR/functions.cpp
@@ -22,7 +22,7 @@ void STR::ConnectFormat(FormatQ *format)
 void STR::slotReadCommand(command_t command)
 {
     qDebug() << "Recieved command from format";
-    qDebug() << "Command: " << command.code << ", pars: " << command.par1 << " " << command.par2;
+    qDebug() << "Command: " << command.code << ", pars: " << command.par1.toInt() << " " << command.par2.toInt();
 
     if(bServerNode) {
